pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "workman"
version = "0.1.0"
description = "Domain operation -> Storacle plan compiler"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["jsonschema>=4.0.0", "python-ulid>=2.0.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/workman"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
